+++ CHANGELOG.md
@@ -521,6 +521,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Recursion
 
+## 0118 - Pascal's Triangle
+
+- Nothing
+
 ## 0415 - Add strings
 
 - Full adder